requirements/base.txt
Django<1.9
python-dateutil